# agent_api_http/openapi-full.yaml
openapi: 3.1.0
info:
  title: UniCore HTTP API
  description: ''
  license:
    name: Apache 2.0
  version: 0.0.0-semantically-released
servers:
- url: http://localhost:3033
  description: Local development
paths:
...
  /request/{request_id}:
    get:
      tags:
      - OID4VP / SIOPv2
      - Protocol
      summary: Get a request object
      description: |-
        Instead of directly embedding the Authorization Request into a QR-code or deeplink, the `Relying Party` can embed a
        `request_uri` that points to this endpoint from where the Authorization Request Object can be retrieved, as
        described in [RFC 9101](https://www.rfc-editor.org/rfc/rfc9101.html#name-passing-a-request-object-by-).
      operationId: request_object
      parameters:
      - name: request_id
        in: path
        description: Authorization request ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Signed authorization request object
          content:
            application/oauth-authz-req+jwt:
              schema:
                type: string
        '404':
          description: The authorization request does not exist
        '500':
          description: The authorization request could not be retrieved
...
components:
...
  securitySchemes:
    access_token:
      type: http
      scheme: bearer
      description: Access token issued by the `/auth/token` endpoint.
...